// lib/api/models/SubjectiveDisplayStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Derived, never stored. EXPIRED means started but not submitted before the effective end time.
 */
export enum SubjectiveDisplayStatus {
    NOT_STARTED = 'NOT_STARTED',
    IN_PROGRESS = 'IN_PROGRESS',
    SUBMITTED = 'SUBMITTED',
    CHECKED = 'CHECKED',
    EXPIRED = 'EXPIRED',
}